move `HouseholdRecord`
